P4ADEV-1448 moved all inside anpr package
